[Gtk] Controls.Core.csproj: Handlers/Items: extract abstract class LayoutManagerBase
