+++ src/Controls/src/Core/Handlers/Items/Gtk/UIExtensions.NUI/LinearLayoutManager.cs
@@ -11,24 +11,8 @@ namespace Gtk.UIExtensions.NUI
 	/// <summary>
 	/// A <see cref="ICollectionViewLayoutManager"/> implementation which provides linear layout
 	/// </summary>
-	public class LinearLayoutManager : ICollectionViewLayoutManager
+	public class LinearLayoutManager : LayoutManagerBase
 	{
-
-		Size _allocatedSize;
-		bool _isLayouting;
-		Rect _lastLayoutedBound;
-		Dictionary<int, RealizedItem> _realizedItem = new Dictionary<int, RealizedItem>();
-		List<double> _itemSizes = new List<double>();
-		List<bool> _cached = new List<bool>();
-		List<double> _accumulatedItemSizes = new List<double>();
-
-		bool _hasUnevenRows;
-		Size _baseItemBound;
-
-		Size _headerSize;
-		View? _header;
-		Size _footerSize;
-		View? _footer;
 
 		/// <summary>
 		/// Initializes a new instance of the <see cref="LinearLayoutManager"/> class.
@@ -49,34 +33,17 @@ namespace Gtk.UIExtensions.NUI
 		/// <param name="isHorizontal">Layout orientation</param>
 		/// <param name="sizingStrategy">Item size measuring strategy</param>
 		/// <param name="itemSpacing">A space size between items</param>
-		public LinearLayoutManager(bool isHorizontal, ItemSizingStrategy sizingStrategy, int itemSpacing)
+		public LinearLayoutManager(bool isHorizontal, ItemSizingStrategy sizingStrategy, int itemSpacing) : base(isHorizontal, sizingStrategy)
 		{
-			IsHorizontal = isHorizontal;
-			_hasUnevenRows = sizingStrategy == ItemSizingStrategy.MeasureAllItems;
 			ItemSpacing = itemSpacing;
 		}
 
 		/// <summary>
-		/// Whether the item is a layout horizontally 
-		/// </summary>
-		public bool IsHorizontal { get; }
-
-		/// <summary>
 		/// A space size between items
 		/// </summary>
-		public double ItemSpacing { get; }
+		public override double ItemSpacing { get; }
 
-		/// <summary>
-		/// CollectionView that interact with layout manager
-		/// </summary>
-		public ICollectionViewController? Controller { get; set; }
-
-		double BaseItemSize
-		{
-			get => IsHorizontal ? BaseItemBound.Width : BaseItemBound.Height;
-		}
-
-		Size BaseItemBound
+		protected override Size BaseItemBound
 		{
 			get
 			{
@@ -97,49 +64,7 @@ namespace Gtk.UIExtensions.NUI
 
 		double ItemHeightConstraint => IsHorizontal ? _allocatedSize.Height : double.PositiveInfinity;
 
-		double FooterSize => IsHorizontal ? _footerSize.Width : _footerSize.Height;
-
-		double HeaderSize => IsHorizontal ? _headerSize.Width : _headerSize.Height;
-
-		double ItemStartPoint
-		{
-			get
-			{
-				var startPoint = HeaderSize;
-
-				if (startPoint > 0)
-				{
-					startPoint += ItemSpacing;
-				}
-
-				return startPoint;
-			}
-		}
-
-		double FooterSizeWithSpacing
-		{
-			get
-			{
-				var size = FooterSize;
-
-				if (size > 0)
-				{
-					size += ItemSpacing;
-				}
-
-				return size;
-			}
-		}
-
-		public void SizeAllocated(Size size)
-		{
-			_allocatedSize = size;
-			InitializeMeasureCache();
-		}
-
-		Size _scrollCanvasSize;
-
-		public Size GetScrollCanvasSize()
+		public override Size GetScrollCanvasSize()
 		{
 			if (Controller!.Count == 0 || _allocatedSize.Width <= 0 || _allocatedSize.Height <= 0)
 			{
@@ -178,7 +103,7 @@ namespace Gtk.UIExtensions.NUI
 			return _scrollCanvasSize;
 		}
 
-		public void LayoutItems(Rect bound, bool force)
+		public override void LayoutItems(Rect bound, bool force)
 		{
 			if (_allocatedSize.Width <= 0 || _allocatedSize.Height <= 0)
 				return;
@@ -226,82 +151,7 @@ namespace Gtk.UIExtensions.NUI
 			_isLayouting = false;
 		}
 
-		public void ItemInserted(int inserted)
-		{
-			var items = _realizedItem.Keys.OrderByDescending(key => key);
-
-			foreach (var index in items)
-			{
-				if (index >= inserted)
-				{
-					_realizedItem[index + 1] = _realizedItem[index];
-				}
-			}
-
-			if (_realizedItem.ContainsKey(inserted))
-			{
-				_realizedItem.Remove(inserted);
-			}
-			else
-			{
-				var last = items.LastOrDefault();
-
-				if (last >= inserted)
-				{
-					_realizedItem.Remove(last);
-				}
-			}
-
-			UpdateInsertedSize(inserted);
-
-			_scrollCanvasSize = new Size(0, 0);
-			Controller!.ContentSizeUpdated();
-		}
-
-		public void ItemRemoved(int removed)
-		{
-			if (_realizedItem.ContainsKey(removed))
-			{
-				Controller!.UnrealizeView(_realizedItem[removed].Holder);
-				_realizedItem.Remove(removed);
-			}
-
-			var items = _realizedItem.Keys.OrderBy(key => key);
-
-			foreach (var index in items)
-			{
-				if (index > removed)
-				{
-					_realizedItem[index - 1] = _realizedItem[index];
-				}
-			}
-
-			var last = items.LastOrDefault();
-
-			if (last > removed)
-			{
-				_realizedItem.Remove(last);
-			}
-
-			UpdateRemovedSize(removed);
-
-			_scrollCanvasSize = new Size(0, 0);
-			Controller!.ContentSizeUpdated();
-		}
-
-		public void ItemUpdated(int index)
-		{
-			if (_realizedItem.ContainsKey(index))
-			{
-				var bound = _realizedItem[index].Holder.GetBounds();
-				Controller!.UnrealizeView(_realizedItem[index].Holder);
-				var view = Controller!.RealizeView(index);
-				_realizedItem[index].Holder = view;
-				view.UpdateBounds(bound);
-			}
-		}
-
-		public Rect GetItemBound(int index)
+		public override Rect GetItemBound(int index)
 		{
 			double itemSize = 0;
 			double startPoint = 0;
@@ -342,44 +192,7 @@ namespace Gtk.UIExtensions.NUI
 				new Rect(0, startPoint, _allocatedSize.Width, itemSize);
 		}
 
-		public void Reset()
-		{
-			foreach (var realizedItem in _realizedItem.Values.ToList())
-			{
-				Controller!.UnrealizeView(realizedItem.Holder);
-			}
-
-			_realizedItem.Clear();
-			_scrollCanvasSize = new Size(0, 0);
-			Controller!.ContentSizeUpdated();
-		}
-
-		public void ItemSourceUpdated()
-		{
-			InitializeMeasureCache();
-		}
-
-		public void ItemMeasureInvalidated(int index)
-		{
-			if (_hasUnevenRows)
-			{
-				if (index >= 0 && _cached.Count > index)
-					_cached[index] = false;
-
-				if (_realizedItem.ContainsKey(index))
-				{
-					Controller!.RequestLayoutItems();
-				}
-			}
-			else if (index == 0)
-			{
-				// Reset item size to measure updated size
-				InitializeMeasureCache();
-				Controller!.RequestLayoutItems();
-			}
-		}
-
-		public int GetVisibleItemIndex(double x, double y)
+		public override int GetVisibleItemIndex(double x, double y)
 		{
 			double coordinate = IsHorizontal ? x : y;
 			double canvasSize = IsHorizontal ? _scrollCanvasSize.Width : _scrollCanvasSize.Height;
@@ -405,125 +218,22 @@ namespace Gtk.UIExtensions.NUI
 			}
 		}
 
-		public double GetScrollBlockSize()
-		{
-			return BaseItemSize + ItemSpacing;
-		}
-
-		public double GetScrollColumnSize()
+		public override double GetScrollColumnSize()
 		{
 			return (IsHorizontal ? BaseItemBound.Height : BaseItemBound.Width);
 		}
 
-		public void SetHeader(View? header, Size size)
-		{
-			bool contentSizeChanged = false;
-
-			if (IsHorizontal)
-			{
-				if (_headerSize.Width != size.Width)
-					contentSizeChanged = true;
-			}
-			else
-			{
-				if (_headerSize.Height != size.Height)
-					contentSizeChanged = true;
-			}
-
-			_header = header;
-			_headerSize = size;
-
-			if (contentSizeChanged)
-			{
-				InitializeMeasureCache();
-				Controller!.ContentSizeUpdated();
-			}
-
-			if (_header != null)
-			{
-				var bound = new Rect(0, 0, _headerSize.Width, _headerSize.Height);
-
-				if (IsHorizontal)
-				{
-					bound.Height = _allocatedSize.Height;
-				}
-				else
-				{
-					bound.Width = _allocatedSize.Width;
-				}
-
-				_header.UpdateBounds(bound);
-			}
-		}
-
-		public void SetFooter(View? footer, Size size)
-		{
-			bool contentSizeChanged = false;
-
-			if (IsHorizontal)
-			{
-				if (_footerSize.Width != size.Width)
-					contentSizeChanged = true;
-			}
-			else
-			{
-				if (_footerSize.Height != size.Height)
-					contentSizeChanged = true;
-			}
-
-			_footer = footer;
-			_footerSize = size;
-
-			if (contentSizeChanged)
-			{
-				InitializeMeasureCache();
-				Controller!.ContentSizeUpdated();
-			}
-
-			UpdateFooterPosition();
-		}
-
-		public int NextRowItemIndex(int index)
+		public override int NextRowItemIndex(int index)
 		{
 			return Math.Min(index + 1, Controller!.Count - 1);
 		}
 
-		public int PreviousRowItemIndex(int index)
+		public override int PreviousRowItemIndex(int index)
 		{
 			return Math.Max(index - 1, 0);
 		}
 
-		void UpdateFooterPosition()
-		{
-			if (_footer == null)
-				return;
-
-			var point = new Point();
-
-			if (IsHorizontal)
-			{
-				point.X += (GetScrollCanvasSize().Width - _footerSize.Width);
-			}
-			else
-			{
-				point.Y += (GetScrollCanvasSize().Height - _footerSize.Height);
-			}
-
-			var bound = new Rect(point, _footerSize);
-
-			if (IsHorizontal)
-			{
-				bound.Height = _allocatedSize.Height;
-			}
-			else
-			{
-				bound.Width = _allocatedSize.Width;
-			}
-
-			_footer.UpdateBounds(bound);
-		}
-
-		void InitializeMeasureCache()
+		protected override void InitializeMeasureCache()
 		{
 			_baseItemBound = Size.Zero;
 			_scrollCanvasSize = new Size(0, 0);
@@ -553,12 +263,12 @@ namespace Gtk.UIExtensions.NUI
 			Controller!.ContentSizeUpdated();
 		}
 
-		int GetStartIndex(Rect bound, double itemSize)
+		protected override int GetStartIndex(Rect bound, double itemSize)
 		{
 			return (int)((ViewPortStartPoint(bound) - ItemStartPoint) / itemSize);
 		}
 
-		int GetStartIndex(Rect bound)
+		protected override int GetStartIndex(Rect bound)
 		{
 			if (!_hasUnevenRows)
 			{
@@ -568,12 +278,12 @@ namespace Gtk.UIExtensions.NUI
 			return FindFirstGreaterOrEqualTo(_accumulatedItemSizes, ViewPortStartPoint(bound));
 		}
 
-		int GetEndIndex(Rect bound, double itemSize)
+		protected override int GetEndIndex(Rect bound, double itemSize)
 		{
 			return (int)Math.Ceiling(ViewPortEndPoint(bound) / (double)itemSize) - 1;
 		}
 
-		int GetEndIndex(Rect bound)
+		protected override int GetEndIndex(Rect bound)
 		{
 			if (!_hasUnevenRows)
 			{
@@ -583,112 +293,13 @@ namespace Gtk.UIExtensions.NUI
 			return FindFirstGreaterOrEqualTo(_accumulatedItemSizes, ViewPortEndPoint(bound));
 		}
 
-		double ViewPortStartPoint(Rect viewPort)
+		protected override void BuildAccumulatedSize()
 		{
-			return IsHorizontal ? viewPort.X : viewPort.Y;
-		}
+			_accumulatedItemSizes = new List<double>();
+			int n = _itemSizes.Count;
 
-		double ViewPortEndPoint(Rect viewPort)
-		{
-			return ViewPortStartPoint(viewPort) + ViewPortSize(viewPort);
-		}
-
-		double ViewPortSize(Rect viewPort)
-		{
-			return IsHorizontal ? viewPort.Width : viewPort.Height;
-		}
-
-		void UpdateAccumulatedItemSize(int index, double diff)
-		{
-			for (int i = index; i < _accumulatedItemSizes.Count; i++)
-			{
-				_accumulatedItemSizes[i] += diff;
-			}
-
-			if (_scrollCanvasSize.Width > 0 && _scrollCanvasSize.Height > 0)
-			{
-
-				if (IsHorizontal)
-				{
-					_scrollCanvasSize.Width += diff;
-				}
-				else
-				{
-					_scrollCanvasSize.Height += diff;
-				}
-			}
-
-			UpdateFooterPosition();
-		}
-
-		void UpdateRemovedSize(int removed)
-		{
-			if (!_hasUnevenRows)
-				return;
-
-			var removedSize = _itemSizes[removed];
-			_itemSizes.RemoveAt(removed);
-			UpdateAccumulatedItemSize(removed, -removedSize);
-			_accumulatedItemSizes.RemoveAt(removed);
-			_cached.RemoveAt(removed);
-		}
-
-		void UpdateInsertedSize(int inserted)
-		{
-			if (!_hasUnevenRows)
-				return;
-
-			_cached.Insert(inserted, false);
-			_itemSizes.Insert(inserted, BaseItemSize);
-			_accumulatedItemSizes.Insert(inserted, 0);
-			_accumulatedItemSizes[inserted] = inserted > 0 ? _accumulatedItemSizes[inserted - 1] : ItemStartPoint;
-			UpdateAccumulatedItemSize(inserted, BaseItemSize);
-		}
-
-		bool ShouldRearrange(Rect viewport)
-		{
-			if (_isLayouting)
-				return false;
-
-			if (_lastLayoutedBound.Size != viewport.Size)
-				return true;
-
-			var diff = IsHorizontal ? Math.Abs(_lastLayoutedBound.X - viewport.X) : Math.Abs(_lastLayoutedBound.Y - viewport.Y);
-
-			if (diff > BaseItemSize)
-				return true;
-
-			return false;
-		}
-
-		static int FindFirstGreaterOrEqualTo(IList<double> data, double value)
-		{
-			if (data.Count == 0)
-				return 0;
-
-			int start = 0;
-			int end = data.Count - 1;
-
-			while (start < end)
-			{
-				int mid = (start + end) / 2;
-
-				if (data[mid] < value)
-				{
-					start = mid + 1;
-				}
-				else
-				{
-					end = mid - 1;
-				}
-			}
-
-			if (data[start] < value)
-			{
-				start++;
-			}
-
-			return start;
+			for (int i = 0; i < n; i++)
+			{ }
 		}
 
 	}
