+++ src/Controls/src/Core/Handlers/Items/Gtk/UIExtensions.NUI/RealizedItem.cs
@@ -1,0 +1,16 @@
+namespace Gtk.UIExtensions.NUI;
+
+public class RealizedItem
+{
+
+	public RealizedItem(ViewHolder holder, int index)
+	{
+		Holder = holder;
+		Index = index;
+	}
+
+	public ViewHolder Holder { get; set; }
+
+	public int Index { get; set; }
+
+}
